Moved fly_by_camera to its own mod
