+++ src/main.rs
@@ -12,14 +12,13 @@ use bevy_inspector_egui;
 #[macro_use]
 mod macros;
 
-mod fly_by_camera;
-use fly_by_camera::FlyByCameraPlugin;
+mod camera;
 
 mod debug;
 use debug::DebugPlugin;
 
 mod world;
-use world::WorldPlugin;
+use world::{rendering::LandscapeCenter, WorldPlugin, terraformation::TerraformationCenter};
 
 mod ui;
 use ui::UiPlugin;
@@ -38,7 +37,7 @@ fn main() {
     .insert_resource(ImageSettings::default_nearest())
     .add_plugins(DefaultPlugins)
     .add_plugin(DebugPlugin)
-    .add_plugin(FlyByCameraPlugin)
+    .add_plugin(camera::CameraPlugin)
     .add_plugin(WorldPlugin)
     .add_plugin(UiPlugin)
     .add_startup_system(setup);
@@ -54,9 +53,22 @@ fn setup(
     mut meshes: ResMut<Assets<Mesh>>,
     mut materials: ResMut<Assets<StandardMaterial>>,
 ) {
-    // cube
+    // camera
+    commands
+        .spawn_bundle(Camera3dBundle::default())
+        .insert(camera::MainCamera)
+        .insert(Name::new("Main Camera"))
+        .insert(LandscapeCenter)
+        .insert(TerraformationCenter);
+
+    // focus
     commands.spawn_bundle(PbrBundle {
-        mesh: meshes.add(Mesh::from(shape::Cube { size: 0.1 })),
+        transform: Transform::from_xyz(0.0, 20.0, 0.0),
+        mesh: meshes.add(Mesh::from(shape::Capsule {
+            radius: 0.25,
+            depth: 1.5,
+            ..default()
+        })),
         material: materials.add(Color::rgb(0.3, 0.3, 0.3).into()),
         ..Default::default()
     });
