+++ src/world/debug/mod.rs
@@ -5,7 +5,10 @@ use bevy::{
     render::mesh::{Indices, PrimitiveTopology},
 };
 
-use crate::{fly_by_camera::FlyByCamera, world::rendering::*};
+use crate::{
+    camera::{self, MainCamera},
+    world::rendering::*,
+};
 use projekto_core::*;
 use projekto_shaping as shaping;
 
@@ -23,15 +26,19 @@ impl Plugin for WireframeDebugPlugin {
             .add_startup_system(setup_wireframe_shader)
             .add_asset::<WireframeMaterial>()
             .add_plugin(MaterialPlugin::<WireframeMaterial>::default())
+            .add_system_set(
+                SystemSet::new()
+                    .with_run_criteria(camera::fly_by::is_active)
+                    .with_system(do_raycast)
+                    .with_system(remove_voxel)
+                    .with_system(add_voxel),
+            )
             .add_system(toggle_mesh_wireframe)
             .add_system(toggle_chunk_voxels_wireframe)
             .add_system(toggle_landscape_pause)
             .add_system(draw_voxels)
-            .add_system(do_raycast)
             .add_system(draw_raycast)
-            .add_system(check_raycast_intersections)
-            .add_system(remove_voxel)
-            .add_system(add_voxel);
+            .add_system(check_raycast_intersections);
     }
 }
 
@@ -317,17 +324,13 @@ fn generate_voxel_edges_mesh(voxels: &[IVec3]) -> (Vec<[f32; 3]>, Vec<u32>) {
 fn do_raycast(
     mut commands: Commands,
     keyboard: Res<Input<KeyCode>>,
-    q_cam: Query<(&Transform, &FlyByCamera)>,
+    q_cam: Query<&Transform, With<MainCamera>>,
 ) {
     if !keyboard.just_pressed(KeyCode::F3) {
         return;
     }
 
-    if let Ok((transform, camera)) = q_cam.get_single() {
-        if !camera.active {
-            return;
-        }
-
+    if let Ok(transform) = q_cam.get_single() {
         let raycast = RaycastDebug {
             origin: transform.translation,
             dir: transform.rotation.mul_vec3(Vec3::Z).normalize() * -1.0,
@@ -472,7 +475,7 @@ fn draw_raycast(
 }
 
 fn remove_voxel(
-    q_cam: Query<(&Transform, &FlyByCamera)>,
+    q_cam: Query<&Transform, With<MainCamera>>,
     mouse_input: Res<Input<MouseButton>>,
     mut set_voxel_writer: EventWriter<CmdChunkUpdate>,
     mut query: ChunkSystemRaycast,
@@ -507,11 +510,7 @@ fn remove_voxel(
             return;
         }
 
-        if let Ok((transform, camera)) = q_cam.get_single() {
-            if !camera.active {
-                return;
-            }
-
+        if let Ok(transform) = q_cam.get_single() {
             let origin = transform.translation;
             let dir = transform.rotation.mul_vec3(Vec3::Z).normalize() * -1.0;
             let range = 100.0;
@@ -522,7 +521,7 @@ fn remove_voxel(
 }
 
 fn add_voxel(
-    q_cam: Query<(&Transform, &FlyByCamera)>,
+    q_cam: Query<&Transform, With<MainCamera>>,
     mouse_input: Res<Input<MouseButton>>,
     mut set_voxel_writer: EventWriter<CmdChunkUpdate>,
     mut query: ChunkSystemRaycast,
@@ -564,11 +563,7 @@ fn add_voxel(
             return;
         }
 
-        if let Ok((transform, camera)) = q_cam.get_single() {
-            if !camera.active {
-                return;
-            }
-
+        if let Ok(transform) = q_cam.get_single() {
             let origin = transform.translation;
             let dir = transform.rotation.mul_vec3(Vec3::Z).normalize() * -1.0;
             let range = 100.0;
